# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIRAMA API",
        "description": "API untuk aplikasi mobile SIRAMA (Sistem Informasi Administrasi Kependudukan). Autentikasi menggunakan Laravel Sanctum Bearer Token.",
        "contact": {
            "email": "admin@sirama.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sirama.tunggulmajid.my.id/api",
            "description": "SIRAMA API Server (Production)"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "SIRAMA API Server (Local)"
        }
    ],
    "paths": {
...
        "/pengajuan/akta-kematian": {
            "post": {
                "tags": [
                    "Pengajuan - Akta Kematian"
                ],
                "summary": "Buat pengajuan Akta Kematian baru",
                "description": "Membuat pengajuan Akta Kematian sekaligus mengisi form dan mengupload semua dokumen dalam satu request (multipart/form-data).",
                "operationId": "536eaf8bd713367984e2d4b4e464498e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_lengkap_anggota",
                                    "alamat_lengkap_anggota",
                                    "nik_anggota",
                                    "file_kk_asli",
                                    "file_ktp_asli",
                                    "file_sk_kematian"
                                ],
                                "properties": {
                                    "nama_lengkap_anggota": {
                                        "type": "string",
                                        "example": "Siti Aminah"
                                    },
                                    "alamat_lengkap_anggota": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1 RT 001 RW 002"
                                    },
                                    "nik_anggota": {
                                        "type": "string",
                                        "example": "3277010101500002"
                                    },
                                    "file_kk_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_sk_kematian": {
                                        "description": "Surat keterangan kematian",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan Akta Kematian berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pengajuan/akta-kematian/{id}": {
            "post": {
                "tags": [
                    "Pengajuan - Akta Kematian"
                ],
                "summary": "Update pengajuan Akta Kematian",
                "operationId": "96b03c54ca863eb3f1c779225efeffaf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "nik": {
                                        "type": "string"
                                    },
                                    "no_whatsapp": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "enum": [
                                            "L",
                                            "P"
                                        ]
                                    },
                                    "pekerjaan": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "desa": {
                                        "type": "string"
                                    },
                                    "rt": {
                                        "type": "string"
                                    },
                                    "rw": {
                                        "type": "string"
                                    },
                                    "nama_lengkap_anggota": {
                                        "type": "string"
                                    },
                                    "alamat_lengkap_anggota": {
                                        "type": "string"
                                    },
                                    "nik_anggota": {
                                        "type": "string"
                                    },
                                    "file_kk_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_ktp_asli": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "file_sk_kematian": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengajuan Akta Kematian berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum. Contoh: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}